More test improvements
