+++ test/tokenizer/basic.tokenizer.test.tsx
@@ -3,34 +3,9 @@ import _ from 'lodash';
 import { mount } from 'enzyme';
 import Tokenizer from '../../src/Tokenizer';
 import Keyevent from '../../src/keyevent';
-import { Option } from '../../src/types';
 import simulateTextInput from '../helpers/simulateTextInput';
 import simulateKeyEvent from '../helpers/simulateKeyEvent';
-
-const BEATLES = ['John', 'Paul', 'George', 'Ringo'];
-
-const BEATLES_COMPLEX: Option[] = [
-  {
-    firstName: 'John',
-    lastName: 'Lennon',
-    nameWithTitle: 'John Winston Ono Lennon MBE',
-  },
-  {
-    firstName: 'Paul',
-    lastName: 'McCartney',
-    nameWithTitle: 'Sir James Paul McCartney MBE',
-  },
-  {
-    firstName: 'George',
-    lastName: 'Harrison',
-    nameWithTitle: 'George Harrison MBE',
-  },
-  {
-    firstName: 'Ringo',
-    lastName: 'Starr',
-    nameWithTitle: 'Richard Starkey Jr. MBE',
-  },
-];
+import { BEATLES, BEATLES_COMPLEX } from '../helpers/data';
 
 describe('TypeaheadTokenizer Component', () => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
